pyproject.toml
[project]
name = "aqi-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fancyimpute>=0.7.0",
    "fastapi>=0.115.6",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "holidays>=0.64",
    "ipykernel>=6.29.5",
    "lxml>=5.3.0",
    "matplotlib>=3.10.0",
    "missingno>=0.5.2",
    "numpy==1.26.4",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pmdarima>=2.0.4",
    "prophet>=1.1.6",
    "python-aqi>=0.6.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "statsforecast>=2.0.0",
    "statsmodels>=0.14.4",
    "tensorflow>=2.18.0",
    "uvicorn>=0.34.0",
    "xgboost>=2.1.3",
]
